requirements.txt
websocket-client>=1.6.0
torch>=2.0.0
numpy>=1.24.0
requests>=2.31.0
psutil>=5.9.0
coremltools>=7.0

